Trying to fix MSVC warning

+++ modules/dsp/chowdsp_filters/Other/chowdsp_FIRPolyphaseDecimator.h
@@ -35,7 +35,7 @@ public:
         {
             buffers.emplace_back (numChannels, maxBlockSizeIn / decimationFactor);
 
-            auto& filter = filters.emplace_back (coeffsPerFilter);
+            auto& filter = filters.emplace_back ((int) coeffsPerFilter);
             filter.prepare (numChannels);
 
             std::fill (oneFilterCoeffs.begin(), oneFilterCoeffs.end(), T {});
